requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
python-dateutil==2.8.2
